Fix moon fetch: use sunrise-sunset.org/v2 plain HTTP (sunrisesunset.io is HTTPS-only)
- sunrisesunset.io 301-redirects HTTP->HTTPS, so WiFiClient got the redirect
  body and failed to parse; switch to api.sunrise-sunset.org/v2 which explicitly
  supports non-TLS for ESP8266 and returns full sun+moon JSON over plain HTTP
- Parse ISO 8601 local times (slice HH:MM at offset 11/14); null moonset -> empty
- Derive glyph phase fraction from moon_phase name + illumination (no moon_phase_value)
- Clean unused pbPort (netmon) and signedness warnings (portal monitor parse)
- README updated (data source + troubleshooting)

+++ src/portal.cpp
@@ -146,8 +146,8 @@ static void handle_save() {
     m.replace(" ", "");
     for (int i = 0; i < MONITOR_MAX; i++) cfg.monitors[i][0] = 0;
     int idx = 0, start = 0;
-    for (int i = 0; i <= m.length() && idx < MONITOR_MAX; i++) {
-      if (i == m.length() || m.charAt(i) == ',') {
+    for (unsigned int i = 0; i <= (unsigned int)m.length() && idx < MONITOR_MAX; i++) {
+      if (i == (unsigned int)m.length() || m.charAt(i) == ',') {
         String host = m.substring(start, i);
         start = i + 1;
         if (host.length() > 0 && host.length() < MONITOR_LEN) {
@@ -243,8 +243,8 @@ void portal_begin() {
   m.replace(" ", "");
   for (int i = 0; i < MONITOR_MAX; i++) cfg.monitors[i][0] = 0;
   int idx = 0, start = 0;
-  for (int i = 0; i <= m.length() && idx < MONITOR_MAX; i++) {
-    if (i == m.length() || m.charAt(i) == ',') {
+  for (unsigned int i = 0; i <= (unsigned int)m.length() && idx < MONITOR_MAX; i++) {
+    if (i == (unsigned int)m.length() || m.charAt(i) == ',') {
       String host = m.substring(start, i);
       start = i + 1;
       if (host.length() > 0 && host.length() < MONITOR_LEN) {
